requirement.txt
a2wsgi==1.10.10
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-cors==0.8.1
aiosignal==1.4.0
aiosmtplib==5.1.0
aiosqlite==0.21.0
alembic==1.18.4
altair==6.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
apache-airflow==3.2.1
apache-airflow-core==3.2.1
apache-airflow-providers-common-compat==1.14.3
apache-airflow-providers-common-io==1.7.2
apache-airflow-providers-common-sql==1.34.0
apache-airflow-providers-fab==3.6.1
apache-airflow-providers-google==21.1.0
apache-airflow-providers-smtp==2.4.5
apache-airflow-providers-standard==1.12.3
apache-airflow-task-sdk==1.2.1
apispec==6.10.0
argcomplete==3.6.3
asgiref==3.11.1
attrs==26.1.0
babel==2.18.0
backoff==2.2.1
blinker==1.9.0
cachelib==0.13.0
cachetools==7.0.6
cadwyn==6.2.2
cattrs==26.1.0
certifi==2026.4.22
cffi==2.0.0
chardet==6.0.0.post1
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
colorful==0.5.8
colorlog==6.10.1
cron_descriptor==2.0.8
croniter==6.2.2
cryptography==46.0.7
db-dtypes==1.5.1
decorator==5.2.1
Deprecated==1.3.1
dill==0.4.1
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.18.0
dotenv==0.9.9
email-validator==2.3.0
fastapi==0.136.0
fastapi-cli==0.0.24
fastuuid==0.14.0
filelock==3.29.0
Flask==3.1.3
flask-appbuilder==5.2.0
flask-babel==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.3.0
frozenlist==1.8.0
fsspec==2026.3.0
gcloud-aio-auth==5.4.4
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.6.4
gcsfs==2026.3.0
gitdb==4.0.12
GitPython==3.1.47
google-ads==30.1.0
google-analytics-admin==0.28.0
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-cloud-aiplatform==1.148.1
google-cloud-alloydb==0.9.0
google-cloud-appengine-logging==1.9.0
google-cloud-audit-log==0.5.0
google-cloud-automl==2.19.0
google-cloud-batch==0.21.0
google-cloud-bigquery==3.41.0
google-cloud-bigquery-datatransfer==3.22.0
google-cloud-bigquery-storage==2.37.0
google-cloud-bigtable==2.36.0
google-cloud-build==3.36.0
google-cloud-compute==1.47.0
google-cloud-container==2.64.0
google-cloud-core==2.5.1
google-cloud-datacatalog==3.30.0
google-cloud-dataflow-client==0.13.0
google-cloud-dataform==0.10.0
google-cloud-dataplex==2.18.0
google-cloud-dataproc==5.27.0
google-cloud-dataproc-metastore==1.22.0
google-cloud-dlp==3.36.0
google-cloud-kms==3.12.0
google-cloud-language==2.20.0
google-cloud-logging==3.15.0
google-cloud-managedkafka==0.4.0
google-cloud-memcache==1.15.0
google-cloud-monitoring==2.30.0
google-cloud-orchestration-airflow==1.20.0
google-cloud-os-login==2.20.0
google-cloud-pubsub==2.37.0
google-cloud-redis==2.21.0
google-cloud-resource-manager==1.17.0
google-cloud-run==0.16.0
google-cloud-secret-manager==2.27.0
google-cloud-spanner==3.65.0
google-cloud-speech==2.38.0
google-cloud-storage==3.10.1
google-cloud-storage-control==1.11.0
google-cloud-storage-transfer==1.20.0
google-cloud-tasks==2.22.0
google-cloud-texttospeech==2.36.0
google-cloud-translate==3.26.0
google-cloud-videointelligence==2.19.0
google-cloud-vision==3.13.0
google-cloud-workflows==1.21.0
google-crc32c==1.8.0
google-genai==1.73.1
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
greenback==1.3.0
greenlet==3.4.0
grpc-google-iam-v1==0.14.4
grpc-interceptor==0.15.4
grpcio==1.80.0
grpcio-gcp==0.2.2
grpcio-status==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.12.0
idna==3.13
immutabledict==4.3.1
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
jmespath==1.1.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lance-namespace==0.6.1
lance-namespace-urllib3-client==0.6.1
lazy-object-proxy==1.12.0
libcst==1.8.6
limits==5.8.0
linkify-it-py==2.1.0
litellm==1.82.6
lockfile==0.12.2
looker_sdk==26.6.1
Mako==1.3.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.3.0
marshmallow-sqlalchemy==1.5.0
mdurl==0.1.2
methodtools==0.4.7
mmh3==5.2.1
more-itertools==11.0.2
msgpack==1.1.2
msgspec==0.21.1
multidict==6.7.1
narwhals==2.20.0
natsort==8.4.0
numpy==2.4.4
oauthlib==3.3.1
openai==2.32.0
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-exporter-otlp-proto-http==1.41.0
opentelemetry-exporter-prometheus==0.62b0
opentelemetry-proto==1.41.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==26.1
pandas==2.3.3
pandas-gbq==0.35.0
pathlib_abc==0.5.2
pathspec==1.1.0
pendulum==3.2.0
pillow==12.2.0
platformdirs==4.9.6
pluggy==1.6.0
prison==0.2.1
prometheus_client==0.25.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
psutil==7.2.2
py-spy==0.4.2
pyarrow==24.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic-extra-types==2.11.1
pydantic-settings==2.14.0
pydantic_core==2.46.3
pydata-google-auth==1.9.1
pydeck==0.9.2
Pygments==2.20.0
pygtrie==2.5.0
PyJWT==2.12.1
pylance==4.0.1
pyparsing==3.3.2
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-discovery==1.2.2
python-dotenv==1.2.2
python-multipart==0.0.26
python-slugify==8.0.4
pytz==2026.1.post1
PyYAML==6.0.3
ray==2.55.1
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
rich==13.9.4
rich-argparse==1.7.2
rich-toolkit==0.19.7
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.19.1
scikit-learn==1.8.0
scipy==1.17.1
setproctitle==1.3.7
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smart_open==7.6.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.49
sqlalchemy-bigquery==1.16.0
sqlalchemy-spanner==1.17.3
SQLAlchemy-Utils==0.42.1
sqlparse==0.5.5
starlette==0.52.1
streamlit==1.56.0
structlog==25.5.0
svcs==25.1.0
tabulate==0.10.0
tenacity==9.1.4
termcolor==3.3.0
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typer==0.24.2
types-protobuf==7.34.1.20260408
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uc-micro-py==2.0.0
universal_pathlib==0.3.10
uritemplate==4.2.0
urllib3==2.6.3
uuid6==2025.0.1
uvicorn==0.46.0
uvloop==0.22.1
virtualenv==21.2.4
watchdog==6.0.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.8
wirerope==1.0.0
wrapt==2.1.2
WTForms==3.2.1
yarl==1.23.0
zipp==3.23.1
